Make system installable as pip package (agentmux)
- Rename src/ to agentmux/ to match distribution name
- Move pipeline.py into agentmux/pipeline.py with relative imports
- Add thin pipeline.py shim at repo root for backward compat
- Add pyproject.toml with console script entry point and package data
- Update all test imports from src.* to agentmux.*
- Add test_packaging_requirements.py
- Update CLAUDE.md, README.md, and all docs/ references
