+++ agentmux/prompts/agents/architect.md
