+++ agentmux/prompts/agents/designer.md
