+++ agentmux/prompts/commands/docs.md
