+++ tests/test_tasks_requirements.py
@@ -4,13 +4,13 @@ import tempfile
 import unittest
 from pathlib import Path
 
-from src.prompts import (
+from agentmux.prompts import (
     build_architect_prompt,
     build_change_prompt,
     build_coder_prompt,
     build_reviewer_prompt,
 )
-from src.state import create_feature_files, load_runtime_files
+from agentmux.state import create_feature_files, load_runtime_files
 
 
 class TasksRequirementsTests(unittest.TestCase):
